+++ packages/raw-client/src/createClient.ts
@@ -1,5 +1,6 @@
 import {
   isResponse,
+  JsonRpcRequest,
   type JsonRpcConnection,
   type JsonRpcId,
   type JsonRpcMessage,
@@ -29,6 +30,7 @@ export type ClientRequest<T, TT> = (
 export interface Client {
   disconnect: () => void
   request: ClientRequest<any, any>
+  onNotification: (cb: (req: JsonRpcRequest) => void) => () => void
 }
 
 let nextClientId = 1
@@ -36,6 +38,11 @@ export const createClient = (gProvider: JsonRpcProvider): Client => {
   let clientId = nextClientId++
   const responses = new Map<JsonRpcId, ClientRequestCb<any, any>>()
   const subscriptions = getSubscriptionsManager()
+  const notificationObservers = new Set<(req: JsonRpcRequest) => void>()
+  const onNotification = (cb: (req: JsonRpcRequest) => void) => {
+    notificationObservers.add(cb)
+    return () => notificationObservers.delete(cb)
+  }
 
   let connection: JsonRpcConnection | null = null
   const send = (
@@ -74,9 +81,15 @@ export const createClient = (gProvider: JsonRpcProvider): Client => {
       // at this point, it means that it should be a notification
       const { params } = parsed
       const { subscription: subscriptionId, result, error } = params
-      if (subscriptionId && ("result" in params || error)) {
+      if (
+        subscriptionId &&
+        subscriptions.has(subscriptionId) &&
+        ("result" in params || error)
+      ) {
         if (error) subscriptions.error(subscriptionId, new RpcError(error!))
         else subscriptions.next(subscriptionId, result)
+      } else {
+        notificationObservers.forEach((cb) => cb(parsed))
       }
     } else
       console.warn("Error parsing incomming message: " + JSON.stringify(parsed))
@@ -112,5 +125,6 @@ export const createClient = (gProvider: JsonRpcProvider): Client => {
   return {
     request,
     disconnect,
+    onNotification,
   }
 }